fix go bench
