+++ private_set_intersection/go/BUILD
@@ -17,3 +17,15 @@ go_test(
         "@org_golang_google_protobuf//proto:go_default_library",
     ],
 )
+
+go_test(
+    name = "benchmark_test",
+    srcs = ["benchmark_test.go"],
+    deps = [
+        "//private_set_intersection/go/client",
+        "//private_set_intersection/go/datastructure",
+        "//private_set_intersection/go/server",
+        "//private_set_intersection/proto:psi_go_proto",
+        "@org_golang_google_protobuf//proto:go_default_library",
+    ],
+)
