+++ private_set_intersection/go/server/BUILD
@@ -1,4 +1,4 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "server",
@@ -10,18 +10,6 @@ go_library(
     deps = [
         "//private_set_intersection/go/datastructure",
         "//private_set_intersection/go/version",
-        "//private_set_intersection/proto:psi_go_proto",
-        "@org_golang_google_protobuf//proto:go_default_library",
-    ],
-)
-
-go_test(
-    name = "server_test",
-    srcs = ["server_test.go"],
-    embed = [":server"],
-    deps = [
-        "//private_set_intersection/go/client",
-        "//private_set_intersection/go/datastructure",
         "//private_set_intersection/proto:psi_go_proto",
         "@org_golang_google_protobuf//proto:go_default_library",
     ],
